feat: Move response methods to base class
- Move `send`, `json`,  and `type` methods out of `Response` and into `HTTPBase` class
- Add `jsonBody` alias method to `HTTPBase` class.
