+++ packages/@pollyjs/core/src/-private/response.js
@@ -1,6 +1,5 @@
-import stringify from 'fast-json-stable-stringify';
-import { assert, HTTP_STATUS_CODES } from '@pollyjs/utils';
 import HTTPBase from './http-base';
+import { assert, HTTP_STATUS_CODES } from '@pollyjs/utils';
 
 const DEFAULT_STATUS_CODE = 200;
 
@@ -36,59 +35,10 @@ export default class PollyResponse extends HTTPBase {
     return this;
   }
 
-  type(type) {
-    return this.setHeader('Content-Type', type);
-  }
-
-  send(data) {
-    let body = data;
-
-    switch (typeof body) {
-      case 'string':
-        // String defaulting to html
-        if (!this.hasHeader('Content-Type')) {
-          this.type('text/html');
-        }
-
-        break;
-      case 'boolean':
-      case 'number':
-      case 'object':
-        if (body === null) {
-          body = '';
-        } else {
-          return this.json(body);
-        }
-
-        break;
-    }
-
-    if (typeof body === 'string') {
-      const contentType = this.getHeader('Content-Type');
-
-      // Write strings in utf-8
-      if (contentType && !contentType.includes('charset')) {
-        this.type(`${contentType}; charset=utf-8`);
-      }
-    }
-
-    this.body = body;
-
-    return this;
-  }
-
   sendStatus(status) {
     this.status(status);
     this.type('text/plain');
 
     return this.send(status);
-  }
-
-  json(obj) {
-    if (!this.hasHeader('Content-Type')) {
-      this.type('application/json');
-    }
-
-    return this.send(stringify(obj));
   }
 }
