+++ packages/@pollyjs/core/tests/unit/-private/http-base-test.js
@@ -1,4 +1,5 @@
 import HTTPBase from '../../../src/-private/http-base';
+import stringify from 'fast-json-stable-stringify';
 
 let base;
 
@@ -69,6 +70,86 @@ describe('Unit | HTTPBase', function() {
       expect(headers).to.deep.equal({ one: '1', two: '2' });
     });
 
+    it('.type()', function() {
+      base.type('text/plain');
+      expect(base.getHeader('Content-Type')).to.equal('text/plain');
+    });
+
+    it('.send() - string', function() {
+      base.send('foo');
+
+      expect(base.body).to.equal('foo');
+      expect(base.getHeader('Content-Type')).to.equal(
+        'text/html; charset=utf-8'
+      );
+    });
+
+    it('.send() - boolean, number, & object', function() {
+      [true, 200, {}].forEach(body => {
+        base.type();
+        base.send(body);
+
+        expect(base.body).to.equal(stringify(body));
+        expect(base.getHeader('Content-Type')).to.equal(
+          'application/json; charset=utf-8'
+        );
+      });
+    });
+
+    it('.send() - null & undefined', function() {
+      base.type();
+      base.send(null);
+      expect(base.body).to.equal('');
+      expect(base.hasHeader('Content-Type')).to.be.false;
+
+      base.type();
+      base.send();
+      expect(base.body).to.be.undefined;
+      expect(base.hasHeader('Content-Type')).to.be.false;
+    });
+
+    it('.send() - should not override existing type', function() {
+      base.type('text/plain; charset=utf-9000');
+      base.send('foo');
+
+      expect(base.body).to.equal('foo');
+      expect(base.getHeader('Content-Type')).to.equal(
+        'text/plain; charset=utf-9000'
+      );
+    });
+
+    it('.json()', function() {
+      const obj = { foo: 'bar' };
+
+      base.json(obj);
+
+      expect(base.body).to.equal(stringify(obj));
+      expect(base.getHeader('Content-Type')).to.equal(
+        'application/json; charset=utf-8'
+      );
+    });
+
+    it('.json() - should not override existing type', function() {
+      const obj = { foo: 'bar' };
+
+      base.type('text/plain; charset=utf-9000');
+      base.send(obj);
+
+      expect(base.body).to.equal(stringify(obj));
+      expect(base.getHeader('Content-Type')).to.equal(
+        'text/plain; charset=utf-9000'
+      );
+    });
+
+    it('.jsonBody()', function() {
+      const obj = { foo: 'bar' };
+
+      expect(() => base.jsonBody()).to.throw(Error);
+
+      base.json(obj);
+      expect(base.jsonBody()).to.deep.equal(obj);
+    });
+
     it('.end()', function() {
       base.setHeader('foo', 'bar');
 
@@ -82,6 +163,9 @@ describe('Unit | HTTPBase', function() {
     it('should be chainable', function() {
       expect(base.setHeader('one', '1')).to.equal(base);
       expect(base.setHeaders({ one: '1' })).to.equal(base);
+      expect(base.type('text/plain')).to.equal(base);
+      expect(base.send('body')).to.equal(base);
+      expect(base.json({ foo: 'bar' })).to.equal(base);
       expect(base.end()).to.equal(base);
     });
   });
