+++ packages/@pollyjs/core/tests/unit/-private/response-test.js
@@ -1,6 +1,5 @@
 import PollyResponse from '../../../src/-private/response';
 
-const { stringify } = JSON;
 let response;
 
 describe('Unit | Response', function() {
@@ -33,83 +32,12 @@ describe('Unit | Response', function() {
       });
     });
 
-    it('.type()', function() {
-      response.type('text/plain');
-      expect(response.getHeader('Content-Type')).to.equal('text/plain');
-    });
-
-    it('.send() - string', function() {
-      response.send('foo');
-
-      expect(response.body).to.equal('foo');
-      expect(response.getHeader('Content-Type')).to.equal(
-        'text/html; charset=utf-8'
-      );
-    });
-
-    it('.send() - boolean, number, & object', function() {
-      [true, 200, {}].forEach(body => {
-        response.type();
-        response.send(body);
-
-        expect(response.body).to.equal(stringify(body));
-        expect(response.getHeader('Content-Type')).to.equal(
-          'application/json; charset=utf-8'
-        );
-      });
-    });
-
-    it('.send() - null & undefined', function() {
-      response.type();
-      response.send(null);
-      expect(response.body).to.equal('');
-      expect(response.hasHeader('Content-Type')).to.be.false;
-
-      response.type();
-      response.send();
-      expect(response.body).to.be.undefined;
-      expect(response.hasHeader('Content-Type')).to.be.false;
-    });
-
-    it('.send() - should not override existing type', function() {
-      response.type('text/plain; charset=utf-9000');
-      response.send('foo');
-
-      expect(response.body).to.equal('foo');
-      expect(response.getHeader('Content-Type')).to.equal(
-        'text/plain; charset=utf-9000'
-      );
-    });
-
     it('.sendStatus()', function() {
       response.sendStatus(200);
 
       expect(response.body).to.equal('200');
       expect(response.getHeader('Content-Type')).to.equal(
         'text/plain; charset=utf-8'
-      );
-    });
-
-    it('.json()', function() {
-      const obj = { foo: 'bar' };
-
-      response.json(obj);
-
-      expect(response.body).to.equal(stringify(obj));
-      expect(response.getHeader('Content-Type')).to.equal(
-        'application/json; charset=utf-8'
-      );
-    });
-
-    it('.json() - should not override existing type', function() {
-      const obj = { foo: 'bar' };
-
-      response.type('text/plain; charset=utf-9000');
-      response.send(obj);
-
-      expect(response.body).to.equal(stringify(obj));
-      expect(response.getHeader('Content-Type')).to.equal(
-        'text/plain; charset=utf-9000'
       );
     });
